Move const variables to firstprivate clause
In GCC < 9, const variables are implied shared and it is an error (!!)
to explicitly make them shared.

In GCC >= 9, const variables need to be explicitly added to the clause.

For compatibility with both compiler eras, const variables can be
moved to a firstprivate clause. The threads don't really need their
own private copy, but it still works.

https://gcc.gnu.org/gcc-9/porting_to.html#ompdatasharing